+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to Scrawler will be documented in this file.
 ### Added
 - Added robots.txt parser block
 - Added support for resolving relative URLs
+- Added CSV file result writer
 - Added literal client configuration provider
 - Added total time displayed once operation is done
 - Added ability to set maximum number of crawled URLs